Fix: 예산 삭제 기능 수정 및 Integer필드 전체  -> BigInteger 로 수정

+++ server/budgets/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from members.models import Member
 
 class Budget(models.Model):
-    value = models.IntegerField()
+    value = models.BigIntegerField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     member = models.ForeignKey(Member, on_delete=models.CASCADE)
